+++ src/models/GeneratorOptions.ts
@@ -1,0 +1,9 @@
+/**
+ * Various options for the toTypescript() function.
+ * 
+ * @export
+ * @interface ToTypescriptOptions
+ */
+export interface ToTypescriptOptions {
+    tabSize: number;
+}