+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+#### Added
+- Classmanager that can modify classes in a document ([#127](https://github.com/buehler/typescript-hero/issues/127))
 
 ## [0.10.1]
 #### Added